backend/requirements.txt
# Django & REST (Django 5.x for Python 3.13+)
Django>=5.0
djangorestframework>=3.14.0
django-cors-headers==4.3.1
django-filter>=24.0

# PostgreSQL (Supabase, etc.)
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Payments - PayTabs (Saudi) only
requests==2.31.0

# Notifications
twilio==9.0.4
python-dotenv==1.0.1

# Email
django-ses==3.5.2
sendgrid==6.11.0

# PDF Invoice Generation
reportlab==4.2.0

# Utilities
python-dateutil==2.8.2
phonenumbers==8.13.27
Pillow>=10.0.0
